+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: mcesar-d <mcesar-d@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/05 19:26:15 by mcesar-d          #+#    #+#             */
-/*   Updated: 2022/04/17 05:21:07 by mcesar-d         ###   ########.fr       */
+/*   Updated: 2022/04/28 00:50:56 by mcesar-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *s)
 {
 	char	*new_src;
-	int		index;
+	size_t	index;
 
 	new_src = malloc((ft_strlen(s) + 1) * sizeof(char));
 	index = 0;
